logger interceptor logic to Api Logger
